feat: use singleton pattern
